Refactor and modernize datatools/utils interfaces
Split functions into seperate headers by task for clarity and cohesion:

- bit manipulation
- string manipulation
- floating point manipulation
- Path expansion
- return types

Include these in existing utils header to maintain existing interface.

Modernize code using ISO C++ Guidelines.
